Add support for external modules
* Add batch event emission to DSL module
* Add support for external modules
* Satisfy linter (mostly relating to integer conversion)

Modules can now be run remotely, outside the Mir node,
in a dedicated module server.
The Mir node can use a proxy module that connects to this server
over a websocket interface and relays all the incoming events
to the remote server. The server applies the events to its local module
(which implements the actual module logic) and sends back the results.
Those resulting events are then emitted by the proxy module
(which lives in the Mir node).

+++ stdmodules/factory/factory_test.go
@@ -102,7 +102,7 @@ func TestFactoryModule(t *testing.T) {
 		},
 
 		"02 Instantiate many": func(t *testing.T) {
-			for i := 1; i <= 5; i++ {
+			for i := uint64(1); i <= 5; i++ {
 				evOut, err := echoFactory.ApplyEvents(stdtypes.ListOf(stdevents.NewNewSubmodule(
 					echoFactoryID,
 					echoFactoryID.Then(stdtypes.ModuleID(fmt.Sprintf("inst%d", i))),
